🥅: Make ErrorWithErrorCode to an actual class and use as super class of custom errors
